+++ tools/clang/lib/Sema/SemaDXR.cpp
@@ -1257,238 +1257,4 @@ void DiagnoseClosestHitEntry(Sema &S, FunctionDecl *FD, HLSLShaderAttr *Attr) {
   }
   return;
 }
-
-static bool NodeInputIsCompatible(StringRef& typeName,
-                                  DXIL::NodeLaunchType launchType) {
-  return llvm::StringSwitch<bool>(typeName)
-         .Case("DispatchNodeInputRecord",
-               launchType == DXIL::NodeLaunchType::Broadcasting)
-         .Case("RWDispatchNodeInputRecord",
-               launchType == DXIL::NodeLaunchType::Broadcasting)
-         .Case("GroupNodeInputRecords",
-               launchType == DXIL::NodeLaunchType::Coalescing)
-         .Case("RWGroupNodeInputRecords",
-               launchType == DXIL::NodeLaunchType::Coalescing)
-         .Case("EmptyNodeInput",
-               launchType == DXIL::NodeLaunchType::Coalescing)
-         .Case("ThreadNodeInputRecord",
-               launchType == DXIL::NodeLaunchType::Thread)
-         .Case("RWThreadNodeInputRecord",
-               launchType == DXIL::NodeLaunchType::Thread)
-         .Default(false);
-}
-
-void DiagnoseNodeEntry(Sema &S, FunctionDecl *FD, HLSLShaderAttr *Attr) {
-
-  SourceLocation computeLoc = SourceLocation();
-  SourceLocation nodeLoc = SourceLocation();
-  SourceLocation nodeLaunchLoc = SourceLocation();
-  DXIL::NodeLaunchType nodeLaunchType = DXIL::NodeLaunchType::Invalid;
-  unsigned inputCount = 0;
-
-  // a function may be both compute and work-graph node
-  for (auto *pAttr : FD->specific_attrs<HLSLShaderAttr>()) {
-    DXIL::ShaderKind shaderKind = ShaderModel::KindFromFullName(pAttr->getStage());
-    if (shaderKind == DXIL::ShaderKind::Node) {
-      nodeLoc = pAttr->getLocation();
-    } else if (shaderKind == DXIL::ShaderKind::Compute) {
-      computeLoc = pAttr->getLocation();
-    }
-  }
-  // if this isn't a work-graph node we can quit now
-  if (!nodeLoc.isValid())
-    return;
-
-  // save NodeLaunch type for use later
-  if (auto nodeLaunchAttr = FD->getAttr<HLSLNodeLaunchAttr>()) {
-    nodeLaunchType =
-      ShaderModel::NodeLaunchTypeFromName(nodeLaunchAttr->getLaunchType());
-    nodeLaunchLoc = nodeLaunchAttr->getLocation();
-  } else {
-    nodeLaunchType = DXIL::NodeLaunchType::Broadcasting;
-    nodeLaunchLoc = SourceLocation();
-  }
-
-  // If this is both a compute shader and work-graph node, it may only have
-  // broadcasting launch mode
-  if (computeLoc.isValid() &&
-      nodeLaunchType != DXIL::NodeLaunchType::Broadcasting) {
-    S.Diags.Report(nodeLaunchLoc, diag::err_hlsl_compute_launch_compatibility)
-      << FD->getName() << ShaderModel::GetNodeLaunchTypeName(nodeLaunchType);
-    S.Diags.Report(computeLoc, diag::note_defined_here) << "compute";
-  }
-
-  // Check that if a Thread launch node has the NumThreads attribute the
-  // thread group size is (1,1,1)
-  if (nodeLaunchType == DXIL::NodeLaunchType::Thread) {
-    if (auto NumThreads = FD->getAttr<HLSLNumThreadsAttr>()) {
-      if (NumThreads->getX() != 1 || NumThreads->getY() != 1 ||
-          NumThreads->getZ() != 1) {
-        S.Diags.Report(NumThreads->getLocation(),
-                       diag::err_hlsl_wg_thread_launch_group_size)
-          << NumThreads->getRange();
-        // Only output the note if the source location is valid
-        if (nodeLaunchLoc.isValid())
-          S.Diags.Report(nodeLaunchLoc, diag::note_defined_here)
-              << "Launch type";
-      }
-    }
-  } else if (!FD->hasAttr<HLSLNumThreadsAttr>()) {
-    // All other launch types require the NumThreads attribute.
-    S.Diags.Report(FD->getLocation(), diag::err_hlsl_missing_node_attr)
-      << FD->getName() << ShaderModel::GetNodeLaunchTypeName(nodeLaunchType)
-      << "numthreads";
-  }
-
-  auto *nodeDG = FD->getAttr<HLSLNodeDispatchGridAttr>();
-  auto *nodeMDG = FD->getAttr<HLSLNodeMaxDispatchGridAttr>();
-  if (nodeLaunchType != DXIL::NodeLaunchType::Broadcasting) {
-    // NodeDispatchGrid is only valid for Broadcasting nodes
-    if (nodeDG) {
-      S.Diags.Report(nodeDG->getLocation(), diag::err_hlsl_launch_type_attr)
-        << nodeDG->getSpelling()
-        << ShaderModel::GetNodeLaunchTypeName(DXIL::NodeLaunchType::Broadcasting)
-        << nodeDG->getRange();
-      // Only output the note if the source location is valid
-      if (nodeLaunchLoc.isValid())
-          S.Diags.Report(nodeLaunchLoc, diag::note_defined_here)
-              << "Launch type";
-    }
-    // NodeMaxDispatchGrid is only valid for Broadcasting nodes
-    if (nodeMDG) {
-      S.Diags.Report(nodeMDG->getLocation(), diag::err_hlsl_launch_type_attr)
-        << nodeMDG->getSpelling()
-        << ShaderModel::GetNodeLaunchTypeName(DXIL::NodeLaunchType::Broadcasting)
-        << nodeMDG->getRange();
-      // Only output the note if the source location is valid
-      if (nodeLaunchLoc.isValid())
-          S.Diags.Report(nodeLaunchLoc, diag::note_defined_here)
-              << "Launch type";
-    }
-  } else {
-    // A Broadcasting node must have one of NodeDispatchGrid or
-    // NodeMaxDispatchGrid
-    if (!nodeMDG && ! nodeDG)
-      S.Diags.Report(FD->getLocation(),
-        diag::err_hlsl_missing_dispatchgrid_attr) << FD->getName();
-    // NodeDispatchGrid and NodeMaxDispatchGrid may not be used together
-    if (nodeMDG && nodeDG) {
-      S.Diags.Report(nodeMDG->getLocation(),
-                     diag::err_hlsl_incompatible_node_attr)
-        << FD->getName() << nodeMDG->getSpelling() << nodeDG->getSpelling()
-        << nodeMDG->getRange();
-      S.Diags.Report(nodeDG->getLocation(), diag::note_defined_here)
-        << nodeDG->getSpelling();
-    }
-  }
-
-  if (!FD->getReturnType()->isVoidType())
-    S.Diag(FD->getLocation(), diag::err_shader_must_return_void)
-      << Attr->getStage();
-
-  // Check parameter constraints
-  for (unsigned Idx = 0; Idx < FD->getNumParams(); ++Idx) {
-    ParmVarDecl *Param = FD->getParamDecl(Idx);
-
-    // compute is incompatible with node input/output
-    if (computeLoc.isValid() && hlsl::IsHLSLNodeType(Param->getType())) {
-      S.Diags.Report(Param->getLocation(),
-                     diag::err_hlsl_compute_io_compatibility)
-        << FD->getName() << "node input/output" << Param->getSourceRange();
-      S.Diags.Report(computeLoc, diag::note_defined_here) << "compute";
-    }
-
-    // Check any node input is compatible with the node launch type
-    if (hlsl::IsHLSLNodeInputType(Param->getType())) {
-      inputCount++;
-      const RecordType* RT = Param->getType()->getAs<RecordType>();
-      StringRef typeName = RT->getDecl()->getName();
-      if (nodeLaunchType != DXIL::NodeLaunchType::Invalid &&
-          !NodeInputIsCompatible(typeName, nodeLaunchType)) {
-        S.Diags.Report(Param->getLocation(), diag::err_hlsl_wg_input_kind)
-          << typeName << ShaderModel::GetNodeLaunchTypeName(nodeLaunchType)
-          << (static_cast<unsigned>(nodeLaunchType) - 1) << Param->getSourceRange();
-        if (nodeLaunchLoc.isValid())
-          S.Diags.Report(nodeLaunchLoc, diag::note_defined_here)
-            << "Launch type";
-      }
-      if (inputCount > 1)
-        S.Diags.Report(Param->getLocation(),
-                       diag::err_hlsl_too_many_node_inputs)
-          << FD->getName() << Param->getSourceRange();
-    }
-
-    HLSLMaxRecordsSharedWithAttr *ExistingMRSWA =
-        Param->getAttr<HLSLMaxRecordsSharedWithAttr>();
-    if (ExistingMRSWA) {
-      StringRef sharedName = ExistingMRSWA->getName()->getName();
-      unsigned int arg_idx = 0;
-      bool found = false;
-      while (arg_idx < FD->getNumParams()) {
-        const ParmVarDecl *parmDecl = FD->getParamDecl(arg_idx);
-        // validation that MRSW doesn't reference its own parameter is
-        // already done at
-        // SemaHLSL.cpp:ValidateMaxRecordsSharedWithAttributes so we don't
-        // need to check that arg_idx != Idx.
-        if (parmDecl->getName() == sharedName) {
-          // now we need to check that this parameter has an output record type.
-          hlsl::NodeFlags nodeFlags;
-          if (GetHLSLNodeIORecordType(parmDecl, nodeFlags)) {
-            hlsl::NodeIOProperties node(nodeFlags);
-            if (node.Flags.IsOutputNode()) {
-              found = true;
-              break;
-            }
-          }
-        }
-        arg_idx++;
-      }
-
-      if (!found) {
-        S.Diag(ExistingMRSWA->getLocation(),
-               diag::err_hlsl_maxrecordssharedwith_references_invalid_arg);
-      }
-    }
-  }
-  return;
-}
-
-void DiagnoseEntry(Sema &S, FunctionDecl *FD) {
-  auto Attr = FD->getAttr<HLSLShaderAttr>();
-  if (!Attr)
-    return;
-
-  DXIL::ShaderKind Stage = ShaderModel::KindFromFullName(Attr->getStage());
-  switch (Stage) {
-  case DXIL::ShaderKind::Pixel:
-  case DXIL::ShaderKind::Vertex:
-  case DXIL::ShaderKind::Geometry:
-  case DXIL::ShaderKind::Hull:
-  case DXIL::ShaderKind::Domain:
-  case DXIL::ShaderKind::Library:
-  case DXIL::ShaderKind::Mesh:
-  case DXIL::ShaderKind::Amplification:
-  case DXIL::ShaderKind::Invalid:
-    return;
-  case DXIL::ShaderKind::Callable: {
-    return DiagnoseCallableEntry(S, FD, Attr);
-  }
-  case DXIL::ShaderKind::Miss:
-  case DXIL::ShaderKind::AnyHit: {
-    return DiagnoseMissOrAnyHitEntry(S, FD, Attr, Stage);
-  }
-  case DXIL::ShaderKind::RayGeneration:
-  case DXIL::ShaderKind::Intersection: {
-    return DiagnoseRayGenerationOrIntersectionEntry(S, FD, Attr);
-  }
-  case DXIL::ShaderKind::ClosestHit: {
-    return DiagnoseClosestHitEntry(S, FD, Attr);
-  }
-  case DXIL::ShaderKind::Compute:
-  case DXIL::ShaderKind::Node: {
-    // A compute shader may also be a node, so we check it here
-    return DiagnoseNodeEntry(S, FD, Attr);
-  }
-  }
-}
 } // namespace hlsl
